Don't mess with how delete works.
This is more control over deleting messages than we want.  First of
all, gmail replicates messages into the inbox and the "all mail"
folder, so deleting it from the inbox already doesn't truly delete
anything.  Second, putting it into the trash isn't helpful, since
there is already a copy in the "all mail" area.  Third, we're not
syncing the trash folder anyway (since it's trash).  Fourth, this
binding breaks the default mutt behavior anyway, which is to have "D"
bound to delete-pattern.

Thus, I am deleting this macro itself.